refactor(home): Move home-related cubits to home screen directory
Moves the following cubits and their states from the generic `presentation/cubits` directory to a more specific `presentation/screens/home/cubits` directory:
- `CategoriesCubit`
- `OffersCubit`
- `ProductsCubit`
- `StoresCubit`

This refactoring improves code organization by co-locating screen-specific state management logic with the screen itself. Import paths across the home screen feature have been updated accordingly.
