+++ lib/presentation/screens/home/builders/sections/category_tabs_builder.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../../../core/design_system/widgets/chip_category.dart';
-import '../../../../cubits/categories/cubit/categories_cubit.dart';
-import '../../../../cubits/categories/cubit/categories_state.dart';
+import '../../cubits/categories/cubit/categories_cubit.dart';
+import '../../cubits/categories/cubit/categories_state.dart';
 
 Widget buildCategoryTabs() {
   return BlocBuilder<CategoriesCubit, CategoriesState>(
